Create a state variable declaration.
