--- examples/testbuild/OUT_DIR/Counter.ftl
@@ -1,3 +1,0 @@
-Counter-Count-text = {$count}
-Counter-IncrText-text = Increment
-Counter-DecrText-text = Decrement
