+++ examples/testbuild/OUT_DIR/TrafficLight.ftl
@@ -1,0 +1,1 @@
+TrafficLight-SwitchText-text = Next
